Fix Add To Victory
There are two cases where enemies get added to victory. When an effect
places them there and when they are defeated. We need to account for
timing differences to ensure the enemy is both in the victory display
and removed from all relevant in play areas. This splits the message
into two messages in order to correctly synchronize event windows.