Prevent skill test subscriber repetition
In hard/expert the gather the Cultist token would reveal another token.
This however caused the Cultist token to appear twice in the skill test
subscribers and the failure criteria would trigger twice. The naive
solution is just to nub the subscribers since tokens have unique ids and
nothing should be subscribed twice to begin with.

Long term we likely want a better solution than subscribers.