Adjust: ability button (#2275)
* fix: resolustion and prologue  unreadable when use phone

* Adjust layout of level up page

* fix: pick supplies page unusable

* fix: broken scenario-cards UI when play "Scenario III: Threads of Fate" in "The Forgotten Age"

* Adjust: the color of the line between location cards and the action counter

Regardless of the browser themes in use, ensure that the line between the location cards and the action counter remains a bright hue.

* Adjust: ability button

Make ability button easy to click.

* Fix The Night Howls

* Fix Taken Captive

* Fix Rivertown 292

* Fix scenario token pool

* Fix Bum's Rush based on ruling

---------- ---------- --------- --------- ---------
The final sentence of Bum’s Rush is not dependent on the previous
sentence; regardless of if you succeed the test, you can move a
non-Elite enemy with the effect.
---------- ---------- --------- --------- ---------

Feel free to reach out to us if any more questions arise!

Sincerely,

Alex Werner, FFG Game Rules Specialist

* Fixes

* Fixes

* Escape the cage does not need to trigger if no enemies

* Modifier should last longer than upkeep

* Fix Knight of the Inner Circle, Terror of the Stars Guardian

* Copy Fixes, ensure timing on that which has no name

* Fix remaining partners

* Smarter tab selection

* Some button adjustments

---------

Co-authored-by: Matthew Mongeau <halogenandtoast@gmail.com>